Eliminated Level 3 warnings. Setting compile warnings to Level 3. (Level 4 just kills eAthena XD)
git-svn-id: http://eathena.googlecode.com/svn/branches/stable@4184 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/char/inter.c
@@ -621,7 +621,7 @@ int inter_check_length(int fd, int length) {
 		length = RFIFOW(fd,2);
 	}
 
-	if (RFIFOREST(fd) < length)	// パケットが未着
+	if ((int)RFIFOREST(fd) < length)	// パケットが未着
 		return 0;
 
 	return length;
